+++ bms/App/lib/imd/src/imd_drivers.c
@@ -51,7 +51,7 @@ void configure_imd_params(void) {
   data[0] = 0xFE;
   imd_send_request(IMD_CAN_ID_REQUEST, IMD_VOLTAGE_MODE, data, 2);
 
-  // ... 
+  // ...
 }
 
 // By default, IMD_Info_General is sent every 100 ms
@@ -66,11 +66,9 @@ void reset_imd_alarm() {
   imd_send_request(IMD_CAN_ID_REQUEST, IMD_RESET_ALARM, data, 1);
 }
 
-void parse_imd_msg(uint8_t can_id, uint8_t *data){
+void parse_imd_msg(uint32_t can_id, uint8_t *data) {
   memcpy(packet.data.raw, data, 8);
   packet.can_id = can_id;
 }
 
-IMD_Packet_t imd_get_data() {
-  return packet;
-}
+IMD_Packet_t imd_get_data() { return packet; }
